Extract AppSec processor context in separate file
